Proof of concept mikktspace tangents
This will calculate mikktspace tangents for models on compile and save them as .tspace external file next to the bsp. This is far from beeing finished, just a proof of concept for now. Having q3map2 only use/generate this tangent space when setting a new flag would be preferred. The lightstage should also be able to read the transformed tangents, so baked models can be lit correctly when using normal maps. This requires another flag because mikkt space baking has two possible ways of calculating the bitangent. Calculated per vertex or per fragment which makes a difference of the generated normal map.